end of part 4

+++ part4/bloglist/controllers/blogs.js
@@ -1,54 +1,92 @@
-const blogsRouter = require('express').Router();
-const Blog = require('../models/blogs');
+const blogsRouter = require('express').Router()
+const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({});
-  response.json(blogs);
-});
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+  response.json(blogs)
+})
 
 blogsRouter.post('/', async (request, response, next) => {
-  const { title, author, url, likes } = request.body;
+  const { title, author, url, likes } = request.body
 
   if (!title || !url) {
-    return response.status(400).json({ error: 'title or url missing' });
+    return response.status(400).json({ error: 'title or url missing' })
   }
+
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = request.user
 
   const blog = new Blog({
     title,
     author,
     url,
     likes: likes || 0,
-  });
+    user: user._id
+  })
 
   try {
-    const savedBlog = await blog.save();
-    response.status(201).json(savedBlog);
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    const populatedBlog = await Blog.findById(savedBlog._id).populate('user', { username: 1, name: 1 })
+    response.status(201).json(populatedBlog)
   } catch (error) {
-    next(error);
+    next(error)
   }
-});
+})
 
-blogsRouter.delete("/:id", async (request, response) => {
+blogsRouter.delete('/:id', async (request, response) => {
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'unauthorized: you can only delete your own blogs' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
-  const { title, author, url, likes } = request.body;
+  const { title, author, url, likes } = request.body
 
   const blog = {
     title,
     author,
     url,
-    likes,
-  };
+    likes
+  }
 
   try {
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true });
-    response.json(updatedBlog);
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true }).populate('user', { username: 1, name: 1 })
+    response.json(updatedBlog)
   } catch (error) {
-    next(error);
+    next(error)
   }
-});
+})
 
-module.exports = blogsRouter;
+module.exports = blogsRouter
